Change to http protocol for submodule
The page build failed for the `master` branch with the following error:

The submodule registered for `./mkdocs-material` could not be cloned. Make sure it's using https:// and that it's a public repo. For more information, see https://docs.github.com/github/working-with-github-pages/troubleshooting-jekyll-build-errors-for-github-pages-sites#invalid-submodule.

For information on troubleshooting Jekyll see:

  https://docs.github.com/articles/troubleshooting-jekyll-builds
	modified:   .gitmodules